CA*: Final Version w/ Visualization

+++ include/amra/arastar.hpp
@@ -60,7 +60,7 @@ public:
 	int replan(
 		std::vector<int>* solution_path,
 		std::vector<int>* action_ids,
-		int m_weight,
+		float m_weight,
 		int* solution_cost) override;
 
 private:
@@ -104,8 +104,8 @@ private:
 	bool improve_path(
 		const double& start_time,
 		double& elapsed_time,
-		int m_weight);
-	void expand(ARAStarState *state, int hidx, int m_weight);
+		float m_weight);
+	void expand(ARAStarState *state, int hidx, float m_weight);
 	bool is_goal(int state_id);
 
 	unsigned int compute_heuristic(int state_id, int hidx);
